requirements.txt
Django>=3.0.1,<3.1.0
djangorestframework>=3.11.0,<3.12.0
beautifulsoup4>=4.8.2,<5.0.0
requests